requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.6
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
numpy==2.3.2
portalocker==3.2.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
# pywin32==311
qdrant-client==1.15.1
requests==2.31.0
sniffio==1.3.1
starlette==0.41.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.30
#python-Levenshtein==0.25.1

# --- NLP / 유사도 계산 ---
sentence-transformers==3.3.1  # 한국어 벡터 모델
fuzzywuzzy==0.18.0            # 문자열 유사도 계산

# --- Redis ---
redis==5.0.1
hiredis==2.3.2

# --- MySQL ---
pymysql==1.1.1

# --- JWT / 설정 ---
PyJWT>=2.8.0
python-dotenv>=1.0.1

# --- AWS S3 ---
boto3
